天气城市

+++ src/main/java/com/company/project/core/AbstractService.java
@@ -1,6 +1,5 @@
 package com.company.project.core;
 
-
 import org.apache.ibatis.exceptions.TooManyResultsException;
 import org.springframework.beans.factory.annotation.Autowired;
 import tk.mybatis.mapper.entity.Condition;
@@ -14,63 +13,63 @@ import java.util.List;
  */
 public abstract class AbstractService<T> implements Service<T> {
 
-    @Autowired
-    protected Mapper<T> mapper;
+	@Autowired
+	protected Mapper<T> mapper;
 
-    private Class<T> modelClass;    // 当前泛型真实类型的Class
+	private Class<T> modelClass; // 当前泛型真实类型的Class
 
-    @SuppressWarnings("unchecked")
+	@SuppressWarnings("unchecked")
 	public AbstractService() {
-        ParameterizedType pt = (ParameterizedType) this.getClass().getGenericSuperclass();
-        modelClass = (Class<T>) pt.getActualTypeArguments()[0];
-    }
+		ParameterizedType pt = (ParameterizedType) this.getClass().getGenericSuperclass();
+		modelClass = (Class<T>) pt.getActualTypeArguments()[0];
+	}
 
-    public void save(T model) {
-        mapper.insertSelective(model);
-    }
+	public void save(T model) {
+		mapper.insertSelective(model);
+	}
 
-    public void save(List<T> models) {
-        mapper.insertList(models);
-    }
+	public void save(List<T> models) {
+		mapper.insertList(models);
+	}
 
-    public void deleteById(Long id) {
-        mapper.deleteByPrimaryKey(id);
-    }
+	public void deleteById(Long id) {
+		mapper.deleteByPrimaryKey(id);
+	}
 
-    public void deleteByIds(String ids) {
-        mapper.deleteByIds(ids);
-    }
+	public void deleteByIds(String ids) {
+		mapper.deleteByIds(ids);
+	}
 
-    public void update(T model) {
-        mapper.updateByPrimaryKeySelective(model);
-    }
+	public void update(T model) {
+		mapper.updateByPrimaryKeySelective(model);
+	}
 
-    public T findById(Integer id) {
-        return mapper.selectByPrimaryKey(id);
-    }
+	public T findById(Long id) {
+		return mapper.selectByPrimaryKey(id);
+	}
 
-    @Override
-    public T findBy(String fieldName, Object value) throws TooManyResultsException {
-        try {
-            T model = modelClass.newInstance();
-            Field field = modelClass.getDeclaredField(fieldName);
-            field.setAccessible(true);
-            field.set(model, value);
-            return mapper.selectOne(model);
-        } catch (ReflectiveOperationException e) {
-            throw new ServiceException(e.getMessage(), e);
-        }
-    }
+	@Override
+	public T findBy(String fieldName, Object value) throws TooManyResultsException {
+		try {
+			T model = modelClass.newInstance();
+			Field field = modelClass.getDeclaredField(fieldName);
+			field.setAccessible(true);
+			field.set(model, value);
+			return mapper.selectOne(model);
+		} catch (ReflectiveOperationException e) {
+			throw new ServiceException(e.getMessage(), e);
+		}
+	}
 
-    public List<T> findByIds(String ids) {
-        return mapper.selectByIds(ids);
-    }
+	public List<T> findByIds(String ids) {
+		return mapper.selectByIds(ids);
+	}
 
-    public List<T> findByCondition(Condition condition) {
-        return mapper.selectByCondition(condition);
-    }
+	public List<T> findByCondition(Condition condition) {
+		return mapper.selectByCondition(condition);
+	}
 
-    public List<T> findAll() {
-        return mapper.selectAll();
-    }
+	public List<T> findAll() {
+		return mapper.selectAll();
+	}
 }
